refactor(webapp): move the chat session store out of the ui layer
Remove lower-layer dependencies on UI code while preserving saved-session compatibility.
