+++ packages/webapp/tests/ui/session-store.test.ts
@@ -4,7 +4,7 @@
 
 import { beforeEach, describe, expect, it } from 'vitest';
 import 'fake-indexeddb/auto';
-import { SessionStore } from '../../src/ui/session-store.js';
+import { SessionStore } from '../../src/scoops/chat-session-store.js';
 import type { ChatMessage, Session } from '../../src/ui/types.js';
 
 describe('SessionStore', () => {
